+++ src/main/java/com/verros/ds/management/DatabaseManagement.java
@@ -3,6 +3,7 @@ package com.verros.ds.management;
 import com.verros.ds.dao.CheckDigitDao;
 import com.verros.ds.entities.CheckDigitJpo;
 import com.verros.rest.dto.CheckDigitDto;
+import com.verros.rest.exception.BusinessException;
 
 import javax.inject.Inject;
 import java.util.List;
@@ -13,18 +14,23 @@ public class DatabaseManagement {
   @Inject
   CheckDigitDao checkDigitDao;
 
-  public void add(String digit) {
+  public long add(String digit) {
     CheckDigitJpo checkDigitJpo = new CheckDigitJpo();
     checkDigitJpo.setDigit(digit);
-    checkDigitDao.add(checkDigitJpo);
+    return checkDigitDao.add(checkDigitJpo);
   }
 
   public List<CheckDigitJpo> getAll(){
     return checkDigitDao.getAll(CheckDigitJpo.class);
   }
 
-  public void delete(Integer id){
-    checkDigitDao.delete(id);
+  public void delete(Integer id) throws BusinessException {
+    if(id>0){
+      checkDigitDao.delete(id);
+    }else{
+      throw new BusinessException("No valid id","BAD_REQUEST");
+    }
+
   }
 
   public void update(CheckDigitDto checkDigitDto){
